# util: decode bytes inplace