--- a/src/storage/types.rs
+++ b/src/storage/types.rs
@@ -68,8 +68,9 @@ impl Key {
     /// Gets and moves the raw representation of this key.
     #[inline]
     pub fn take_raw(self) -> Result<Vec<u8>, codec::Error> {
-        let k = self.0;
-        bytes::decode_bytes_in_place(k, false)
+        let mut k = self.0;
+        bytes::decode_bytes_in_place(&mut k, false)?;
+        Ok(k)
     }
 
     /// Gets the raw representation of this key.