## What have you changed? (mandatory)
- add `decode_bytes_inplace()` to save mallocation
- add `take_raw()` for `Key` which use `decode_bytes_inplace()`
- call `take_raw()` instead of `raw()` in some code place

## What are the type of the changes? (mandatory)
- Improvement (non-breaking change which is an improvement to an existing feature)

## How has this PR been tested? (mandatory)
unit-test

## Does this PR affect documentation (docs/docs-cn) update? (mandatory)
No

## Does this PR affect tidb-ansible update? (mandatory)
No

## Benchmark result if necessary (optional)
I write a micro bench for two decode_bytes_inplace implementation.
one for copying bytes one by one
```rust
 for i in write_offset..(write_offset + ENC_GROUP_SIZE){
     data[i] = data[read_offset];
     read_offset += 1;
}
```
the other for memmoving 8 bytes in one shot
```rust
ptr::copy(
    data.as_ptr().offset(read_offset as isize),
    data.as_mut_ptr().offset(write_offset as isize),
    ENC_GROUP_SIZE - pad_size
);
```
here is the result
```
test util::codec::bytes::tests::bench_decode                 ... bench:   3,291,861 ns/iter (+/- 234,919)
test util::codec::bytes::tests::bench_decode_inplace_copy    ... bench:   4,772,352 ns/iter (+/- 1,063,849)
test util::codec::bytes::tests::bench_decode_inplace_memmove ... bench:   2,591,718 ns/iter (+/- 557,794)
```