@breeswish 2018-08-10
What's the difference between decode_bytes_in_place_memmove and decode_bytes_in_place? BTW I think it should accept a &mut so that it is "in place".