@Connor1996 2018-08-10
@breeswish These are two implementation and you can see the bench result for the difference. I'm working on cleaning this and add some comment.