@Connor1996 2018-08-10
> BTW I think it should accept a &mut so that it is "in place".
if this function return Err, &mut may let the vec incomplete decode and caller may misuse the vec