@breeswish 2018-08-10
Interesting.. ptr::copy is faster than directly copy as u64:

```rust
        unsafe {
            let v = *(&data[read_offset] as *const u8 as *const u64);
            *(&mut data[write_offset] as *mut u8 as *mut u64) = v;
        }
```

@zhangjinpeng1987