@breeswish 2018-08-10
I found out the reason of slow is because of the boundary check in [] operator.