@breeswish 2018-08-10
I successfully made some even more improvements over your code, 30% faster:

```rust
pub fn decode_bytes_in_place4(data: &mut Vec<u8>) -> Result<(), &'static str> {
    unsafe {
        let mut write_offset = 0;
        let mut read_offset = 0;
        let len = data.len();
        if len % 9 != 0 || len == 0 {
            return Err("Unexpected EOF");
        }
        if *data.get_unchecked(len - 1) == 255 {
            // Last byte must be a > 0 padding size.
            return Err("Padding error");
        }
        loop {
            let marker_position = read_offset + ENC_GROUP_SIZE;
            let next_position = marker_position + 1;

            let chunk: u64 = *(data.as_ptr().offset(read_offset as isize) as *const u64);
            *(data.as_mut_ptr().offset(write_offset as isize) as *mut u64) = chunk;

            read_offset = next_position;
            write_offset += ENC_GROUP_SIZE;
            let pad_size = (ENC_MARKER - data.get_unchecked(marker_position)) as usize;
            if pad_size > 0 {
                if pad_size > ENC_GROUP_SIZE {
                    return Err("Padding error");
                }
                if data[marker_position - pad_size..marker_position] != ENC_PADDING[..pad_size] {
                    return Err("Padding error");
                }
                data.set_len(write_offset - pad_size);
                return Ok(());
            }
        }
    }
}
```

(notice that I was using `&'static str` because I was writing a standalone test outside TiKV. You need to change it back to our `Error` type)

Result:

With valid padding check (30% faster than this PR):

```
test tests::bench_decode          ... bench:      48,991 ns/iter (+/- 19,410)
test tests::bench_decode_inplace  ... bench:      15,327 ns/iter (+/- 2,364)
test tests::bench_decode_inplace4 ... bench:      10,678 ns/iter (+/- 2,000)
```

Without valid padding check (42% faster than this PR):

```
test tests::bench_decode          ... bench:      42,859 ns/iter (+/- 14,405)
test tests::bench_decode_inplace  ... bench:      15,235 ns/iter (+/- 2,181)
test tests::bench_decode_inplace4 ... bench:       8,836 ns/iter (+/- 1,554)
```

The core idea is to eliminate `if` statements, for example, removing "desc" helps a lot.