@breeswish 2018-08-10
Use `ptr::write` (write aligned memory) instead of `*xxxx = chunk` can be even more faster (benchmark without valid padding check):

```
test tests::bench_decode_inplace4 ... bench:       8,296 ns/iter (+/- 1,978)
```