@zhangjinpeng1987 2018-08-11
@breeswish There is something wrong with your code, because it is decode in place, so the source and destination can overlap, so should use memmove.