@breeswish 2018-08-11
It reads memory to a variable first, then writes to another memory. I think the overlapping does not matter. Otherwise unit tests will fail.