@Connor1996 2018-08-13
@breeswish 
when I only replace `ptr::copy()` with this, seems that it doesn't make any progress.
```
let chunk: u64 = *(data.as_ptr().offset(read_offset as isize) as *const u64);
            *(data.as_mut_ptr().offset(write_offset as isize) as *mut u64) = chunk;
```