@breeswish 2018-08-13
@Connor1996 As said previously, most improvements are caused by eliminating `if` statements, for example, remove the `desc` option and remove most of the branches in loops.