@breeswish 2018-08-13
Also my benchmark is like:

```rust
    #[bench]
    fn bench_decode_inplace4(b: &mut Bencher) {
        let key = [b'x'; 100000];
        let encoded = encode_bytes(&key);
        b.iter(|| {
            let mut value = test::black_box(encoded.clone());
            test::black_box(decode_bytes_in_place4(&mut value));
        });
    }
```

By moving encode_bytes to outside and using a clone instead, the result will be more relevant.