@Connor1996 2018-08-13
new bench result
```
test util::codec::bytes::tests::bench_decode                                   ... bench:       4,486 ns/iter (+/- 4,504)
test util::codec::bytes::tests::bench_decode_inplace                      ... bench:       1,383 ns/iter (+/- 117)
test util::codec::bytes::tests::bench_decode_inplace_write            ... bench:       1,406 ns/iter (+/- 556)
test util::codec::bytes::tests::bench_decode_inplace_write_small  ... bench:          38 ns/iter (+/- 4)
test util::codec::bytes::tests::bench_decode_inplace_small            ... bench:          36 ns/iter (+/- 2)
test util::codec::bytes::tests::bench_decode_small                          ... bench:          69 ns/iter (+/- 3)
```

@breeswish seems that `ptr::write` doesn't make progress either, the prev result maybe cause by Deviation. And I didn't accept the checking code cause it didn't cover the case that there are some timestamp bytes after encoded bytes.