@breeswish 2018-08-13
Wow. How would timestamped bytes affect the correctness for current functions? I previously thought that timestamps are always truncated before passing to the decode function. I tested my function over current codec test cases and it passed. If this is a scenario we need to consider, please add tests to ensure that.