@Connor1996 2018-08-13
current function decodes from front to last padding bytes and just ignores rest bytes. Actually we call `raw()/take_raw()` after `truncate()`, but it maybe misused, I will add test for it.