@breeswish 2018-08-13
> current function decodes from front to last padding bytes and just ignores rest bytes. Actually we call raw()/take_raw() after truncate(), but it maybe misused, I will add test for it.

Please open an issue to track this misuse. As you can see, this kind of misuse prevents us from using a more efficient implementation 🤣 and it is likely to cause mistakes in future because we know nothing about this when looking at the codec.