@breeswish 2018-08-14
@Connor1996 In test cases there is no assurance for this kind of usage at all. Even after you adds these new test cases I bet my implementation will pass it 🤣 (and fail the code in elsewhere)