@Connor1996 2018-08-14
I have already add the tests, and your implementation will regard it as Unexpected EOF or PaddingError